Buffered enc extent polygons to match buffered features
